+++ semcore/i18n-unplugin/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [0.18.13] - 2024-11-22
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/core` [2.36.1 ~> 2.36.2]).
+
 ## [0.18.12] - 2024-11-08
 
 ### Changed